Moved the different startup components out of the StartupManager and into their own separate StartupComponent classes.
