+++ src/main/java/com/synopsys/integration/alert/component/settings/actions/SettingsGlobalApiAction.java
@@ -43,7 +43,7 @@ import com.synopsys.integration.alert.common.security.EncryptionUtility;
 import com.synopsys.integration.alert.component.settings.descriptor.SettingsDescriptor;
 import com.synopsys.integration.alert.database.api.DefaultUserAccessor;
 import com.synopsys.integration.alert.web.security.authentication.saml.SAMLManager;
-import com.synopsys.integration.alert.workflow.startup.SystemValidator;
+import com.synopsys.integration.alert.workflow.startup.component.SystemValidator;
 
 @Component
 public class SettingsGlobalApiAction extends ApiAction {
