+++ src/main/java/com/synopsys/integration/alert/workflow/startup/component/StartupComponent.java
@@ -6,27 +6,12 @@ import org.slf4j.LoggerFactory;
 public abstract class StartupComponent {
     private final Logger logger = LoggerFactory.getLogger(StartupComponent.class);
 
-    private final Integer priorityWeight;
-    private final String ComponentName;
-
-    public StartupComponent(final Integer priorityWeight, final String componentName) {
-        this.priorityWeight = priorityWeight;
-        ComponentName = componentName;
-    }
-
     public void initializeComponent() {
-        final String runningStartupComponentLog = String.format("Running startup component %s with weight %s", getComponentName(), getPriorityWeight());
+        final String runningStartupComponentLog = String.format("Running startup component: %s", getClass().getSimpleName());
         logger.info(runningStartupComponentLog);
-        runComponent();
+        run();
     }
 
-    public abstract void runComponent();
+    public abstract void run();
 
-    public Integer getPriorityWeight() {
-        return priorityWeight;
-    }
-
-    public String getComponentName() {
-        return ComponentName;
-    }
 }
